Fixes due ot review comments
